Caffaina/pyproject.toml
[tool.poetry]
name = "Caffaina"
version = "0.1.0"
description = "A bot"
authors = ["Calvin"]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.4"
Flask = "^2.2.2"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
